Added functionality do that Tasks can be displayed, but drag and drop function does not work
LO2 is listorganizer2 ...don't worry about it
Currently, dragging and dropping replaces the text of the target with
the text from the source

What we want is for the dragged item to be deleted and placed above the
item. Then, according to the specifications, we need to reoreder it so
the newwest item is at the bottom of the priority it was moved into.